Fix the bug where the user always considered unauthorized when invoking BidUp SignalR method
The access token is sent in the query string when SignalR uses WebSockets or ServerSentEvents due to a limitation in Browser APIs. So I have to hook the OnMessageReceived event in order to allow the JWT authentication handler to read the access token from the query string when a WebSocket or Server-Sent Events request comes in. (for more details: https://youtu.be/mrCxfifTepU?si=Zj6Ghx0aZEVFELPL, https://learn.microsoft.com/en-us/aspnet/core/signalr/authn-and-authz).