Fix bug of returning the datetime in non-UTC format
I save all DateTime as UTC in the database. I then parse the UTC time to local time in the client, but by default the json serializer didn't append a "Z". This commit solves that issue by specifying that the time is already in UTC and will only append the 'Z' at the end of it.